12.05

+++ Assets/Scripts/UI/Customize/Module.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Item", menuName = "ScriptableObjects/Module", order = 1)]
 public class Module : Item
 {
-    public float occupancyPoint;
+    public int occupancyPoint;
 
     public int hp;
     public int power;
